# Laying the Groundwork
Today we'll be tackling a [code kata](https://www.codewars.com/) called
[“Delete occurrences of an element if it occurs more than n
times”](https://www.codewars.com/kata/delete-occurrences-of-an-element-if-it-occurs-more-than-n-times)
(what a catchy name!). The goal of this kata is to implement a function
called `deleteNth`. The function accepts a list of numbers as its first
parameter and another number, `N`, as its second parameter. `deleteNth`
should iterate over each number in the provided list, and remove and
numbers that have appeared more than `N` times before returning the
resulting list.

While this is a fairly simple problem, we're going to solve it in a very
deliberate way in order to practice building better software.

This first commit lays the groundwork for our future work. We've set up
a simple [Node.js](https://nodejs.org/en/) project that uses
[Babel](http://babeljs.io/) for ES6 support and
[Mocha](https://mochajs.org/)/[Chai](http://chaijs.com/) for testing.