# Take What We're Given
One of the challenges of real-world problems is teasing out the best
interface for a given task. Code katas are different from real-world
problems in that we're usually given the interface we're supposed to
implement upfront.

In this case, we know that we need to implement a function called
`deleteNth` which accepts an array of numbers as its first argument
(`arr`), and a number, `N`, as its second parameter (`x`).

Eventually, `deleteNth` will return an array of numbers, but we need to
take this one step at a time.