# Forging Ahead
As we add more test cases, things begin to get more complicated. In our
next test we assert that `deleteNth([1, 2], 1)` should equal `[1, 2]`.
Unfortunately, our initial solution of always returning an empty array
failed in this case.

    expected [] to deeply equal [ 1, 2 ]

We know that all calls to `deleteNth` where `x` is zero should result in
an empty array, so lets add a guard that checks for that case.

If `x` is not zero, we know that our test expects us to return `[1, 2]`
which is being passed in through `arr`. Knowing that, we can bring our
tests back into a green state by just returning `arr`.