# Getting Real
We added a new test case, and things suddenly became very real. Our new
test expects `deleteNth([1, 1, 2], 1)` to return `[1, 2]`. This means
that the second `1` in the input array should be removed in the result.
After adding this test, our test suite groans and slips into a red
state.

It seems that we have to finally begin implementing a "real" solution
for this problem.

Because we want to conditionally remove elements from an array, my mind
initially gravitates to using
[`filter`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/filter).
We replace our final `return` statement with a block that looks like
this:

    return arr.filter((num) => {
      return seenNum <= x;
    });

Our filter function will only pass through values of `arr` that we’ve
seen (`seenNum`) no more than `x` times. After making this change our
test suite expectedly complains about `seenNum` not being defined. Let's
fix that.

To know how many times we've seen a number, we need to keep track of
each number we see as we move through `arr`. My first instinct is to do
this with a simple object acting as a map from the number we seen to the
number of times we’ve seen it:

    let seen = {};

Because `seen[num]` will initially be `undefined` we need to give it a
default value of `0`:

    seen[num] = (seen[num] || 0) + 1;

Our test suite seems happy with this solution and flips back into a
green state.