# Removing the Base Case
While we're on a refactoring kick, we also notice that the entire zero
base case is no longer necessary. If `N` (`x`) is zero, our filter
function will happily drop every number in `arr`, resulting in an empty
array.

We can remove the entire `if` block at the head of our `deleteNth`
function.